requirements.txt
Mezzanine==1.4.7
django-storages==1.1.8
boto==2.9.5
psycopg2==2.5
gunicorn==0.17.4
django-extensions==1.1.1
easy-thumbnails==1.2
django-dajaxice==0.5.5
sorl-thumbnail==11.12
django-configurations==0.8
